agents/worker_agent1/requirements.txt
pika
python-dotenv
langgraph
langchain
langchain-core
